pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gitten"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.59.0",
    "gitpython>=3.1.40",
    "pyperclip>=1.8.2",
]

[project.scripts]
gitten = "gitten.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
